% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{glamour_palette_names}
\alias{glamour_palette_names}
\title{List Available Glamour Palettes}
\usage{
glamour_palette_names()
}
\value{
A character vector of palette names.
}
\description{
List Available Glamour Palettes
}
