% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_all_palettes}
\alias{show_all_palettes}
\title{Preview All Glamour Palettes}
\usage{
show_all_palettes()
}
\value{
Invisibly returns a named list of ggplot objects.
}
\description{
Calls \code{\link[=show_palette]{show_palette()}} for every registered palette.
}
